requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
python-dotenv==1.0.1
Werkzeug==2.3.6
Adafruit-DHT==1.4.0
blinker==1.6.2
build==1.2.1
click==8.1.3
Flask==2.3.2
gunicorn==23.0.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
mod-wsgi==4.9.4
packaging==23.1
pkg_resources==0.0.0
pyproject_hooks==1.0.0
python-dotenv==1.0.1
RPi.GPIO==0.7.1
RPLCD==1.3.1
tomli==2.0.1
Werkzeug==2.3.6
zipp==3.15.0
